Alter database structure
Aggregated data into single table structure:
1) Instead of working with individual data tables the areas are aggregated into a units table that holds census/NHGIS and NHD units.
2) The landcover rasters are now tiled and aggregated into a landcover table.
3) All landcover area per unit tabulations are now contained in a single table.

The resulting units, landcover and tabulations are contained in the default schema and the individual data tables are contained in a new base_data schema.  This structure makes things much easier when working with the data and connecting the database to Django models as only three models are really needed.

Added a printout that gives the database load duration in days/hours/minutes/seconds.

Created a new nlcd package on PyPI and added it to requirements and removed nlcd.py from this project.